# Optimize readSliceBuf