Before:
```
BenchmarkReadSliceBuf
BenchmarkReadSliceBuf/small
BenchmarkReadSliceBuf/small-16         	 3337833	       380.6 ns/op
BenchmarkReadSliceBuf/mid
BenchmarkReadSliceBuf/mid-16           	  297603	      4010 ns/op
BenchmarkReadSliceBuf/large
BenchmarkReadSliceBuf/large-16         	   20037	     60182 ns/op
```

After:
```
BenchmarkReadSliceBuf
BenchmarkReadSliceBuf/small
BenchmarkReadSliceBuf/small-16         	 5951392	       201.1 ns/op
BenchmarkReadSliceBuf/mid
BenchmarkReadSliceBuf/mid-16           	  590236	      2038 ns/op
BenchmarkReadSliceBuf/large
BenchmarkReadSliceBuf/large-16         	   40591	     29739 ns/op
```